fix: clarify converter and ttl handling
